requirements.txt
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.11.8
django-jinja==2.4.1
django-jinja-bootstrap-form==4.0.3
djangorestframework==3.5.3
django-rest-swagger==2.1.1
django-filter==1.1.0 # Needs updating with Django
fabric>=1.13.2
funcy>=1.10,<2.0
handy==0.7
honcho==0.6.4
Jinja2==2.10
psycopg2==2.7.3.2
redis==2.10.6
celery==3.1.17
tqdm==4.11.2
requests==2.18.4
ftptool==0.7.1
django-cacheops==4.0.3
gzip-reader==0.1
pandas==0.18.1
termcolor==1.1.0
numpy==1.11.1
numexpr==2.6.2
scipy==0.16.0
statsmodels==0.8.0
django_pandas==0.4.5
easydict==1.7
# TODO: switch to proper release once it's out
django-datatable-view-compat==0.8.2
boto==2.38.0
rpy2==2.9.1
django-registration==2.0.4
biopython==1.68
mygene==3.0.0
requests-oauthlib==0.8.0
